turn -> turn_actions. extracted dice resolver to dice package
